Use data from loginHandler
