+++ Makefile
@@ -38,11 +38,13 @@ db:
 	docker compose run --rm backend sh -c "python manage.py dbshell"
 
 clean:
-	docker compose down --rmi all
+	docker compose down --remove-orphans --rmi all
 
 fclean: clean
 	docker volume ls -q | grep database-data | xargs --no-run-if-empty docker volume rm
 	docker volume ls -q | grep redis-data | xargs --no-run-if-empty docker volume rm
+	docker volume ls -q | grep django-static | xargs --no-run-if-empty docker volume rm
+	docker volume ls -q | grep django-media | xargs --no-run-if-empty docker volume rm
 
 test:
 	docker compose run --rm backend sh -c "python manage.py test"
