+++ frontend/public/js/scripts/localPong.js
@@ -1,5 +1,5 @@
 
-export function initPongGame() {
+export function initPongGame(routerData) {
     const canvas = document.getElementById('pongCanvas');
     const scoreZone = document.getElementById('scoreZone');
     if (!canvas) {
@@ -10,7 +10,7 @@ export function initPongGame() {
     }
 
     const gameSocket = new WebSocket(
-        'wss://' + location.host + '/ws/game/local/' + localStorage.getItem('user_id') + '/?token=' + localStorage.getItem('authToken')
+        'wss://' + location.host + '/ws/game/local/' + routerData.user.id + '/?token=' + localStorage.getItem('authToken')
     );
 
     let keyPressed = {"ArrowUp": false, "ArrowDown": false, "w": false, "s": false};
@@ -22,6 +22,7 @@ export function initPongGame() {
                 document.addEventListener('keydown', (e) => {
                     if (e.key in keyPressed) {
                         keyPressed[e.key] = true;
+                        e.preventDefault();
                     }
                     if(e.key.repeat) {
                         return;
