+++ frontend/public/js/utils/loginHandler.js
@@ -1,3 +1,6 @@
+import { printError } from "./toastMessage.js";
+import { isTokenExpired } from "./tokenHandler.js";
+
 export async function getUserInfo() {
     let response = await fetch("/api/user/me/", {
         method: "GET",
@@ -14,10 +17,10 @@ export async function getUserInfo() {
 
 export function isLoggedIn() {
     const authToken = localStorage.getItem('authToken');
-    const isLogged = authToken !== null &&
+    if (authToken !== null &&
         authToken !== undefined &&
-        !isTokenExpired(authToken);
-    if (isLogged) {
+        !isTokenExpired(authToken)
+    ) {
         return true;
     }
     console.log('invalid access token, checking refresh token');
@@ -38,34 +41,13 @@ export function isLoggedIn() {
                 if (Object.hasOwn(data, "access")) {
                     localStorage.setItem("authToken", data.access);
                     return true;
-                } else {
-                    return false;
                 }
+                return false;
             })
             .catch((error) => {
-                console.error("Login error:", error);
-                // Handle login error, e.g., show error message
+                printError(error);
             });
     }
     console.log('invalid refresh token, not logged in');
     return false;
 }
-
-function decodeJWT(token) {
-    const base64Url = token.split('.')[1];
-    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-    const payload = decodeURIComponent(window.atob(base64).split('').map(function (c) {
-        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-    }).join(''));
-
-    return JSON.parse(payload);
-}
-
-function isTokenExpired(token) {
-    const decodedToken = decodeJWT(token);
-    const expirationTime = decodedToken.exp;
-    const currentTime = Math.floor(Date.now() / 1000);
-
-    return expirationTime <= currentTime - 30;
-}
-
