+++ frontend/public/js/utils/tokenHandler.js
@@ -1,0 +1,17 @@
+export function decodeJWT(token) {
+    const base64Url = token.split('.')[1];
+    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
+    const payload = decodeURIComponent(window.atob(base64).split('').map(function (c) {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
+    }).join(''));
+
+    return JSON.parse(payload);
+}
+
+export function isTokenExpired(token) {
+    const decodedToken = decodeJWT(token);
+    const expirationTime = decodedToken.exp;
+    const currentTime = Math.floor(Date.now() / 1000);
+
+    return expirationTime <= currentTime - 30;
+}
